feat(ui): add Discord help link to error toasts
Each error toast now shows a small Discord icon and "Get help in our
discord" link pointing to the project's Discord server.
